Fix parsing of string literal in proc macro argument
Commit 056b0b7d09a4ed49e50182df216361397656425c removed the dependency `syn`.

In doing so it introduced a bug where the string literal passed to
`document_features!(feature_label = "...")` is not parsed anymore and
therefore in the generated Markdown we insert a representation of the
string literal as it appears in the source code instead of its content.

For instance, with the call `document_features!(feature_label = r#"foo\"#)`
we insert in the Markdown the text `r#"foo\"#` instead of `foo\`.