Fix documentation having too many `#`
rustdoc used to remove `#` from the the liens starting with `##`, but it
no longer does that for non-rust languages.
As a result, we shoudln't add the `#` otherwise it will appear with too
many `#`, but we still need to re-add it in the test because it was
stripped by the test.

Fixes #22